API_Integration_assignment/requirements.txt
openai
groq
requests
google-generativeai
cohere